pyproject.toml
[tool.poetry]
name = "Kiyanka"
version = "0.1.0"
description = "CLI image processor"
authors = ["Ruslan Simashev"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
pillow = "^11.2.0"
numpy = "<2.0"
rembg = {extras = ["cpu"], version = "^2.0.66"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
